% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cnv}
\alias{cnv}
\title{Copy Number Variation data}
\format{
A data frame with ??? rows and ??? variables
}
\usage{
data(cnv)
}
\description{
A dataset containing the copy number variation data for 163 individuals for ??? genes across ??? sites total
}
\details{
\itemize{
\item ???
}
}
\keyword{datasets}
