% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{construct.priors}
\alias{construct.priors}
\title{construct.priors}
\usage{
construct.priors()
}
\arguments{
\item{r_tau}{(0.001) hyperparameter for residual precision tau}

\item{d_tau}{(0.001) hyperparameter for residual precision tau}

\item{d_pi}{(1) hyperparameter for prob of kth coefficient != 0}

\item{r_pi}{(1) hyperparameter for prob of kth coefficient != 0}

\item{r_gamma}{(0.001) hyperparameter for precision of kth coefficient}

\item{d_gamma}{(0.001) hyperparameter for precision of kth coefficient}
}
\value{
list with the following hyperparamenters accessible by index or param name
}
\description{
constructs the priors for the clinical model. includes basic bound checks
}
\details{
follows from https://arxiv.org/pdf/1811.02962v1.pdf section 2.3
\itemize{
\item (residual precision) tau ~ Gamma(r_tau,d_tau)
\item (prob of kth coefficient != 0) pi_k ~ Beta(d_pi, r_pi)
\item (precision of kth coefficient) gamma_k ~ Gamma(r_gamma,d_gamma)
Note: all parameters must be positive (>0)
}
}
