% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{data.collapse}
\alias{data.collapse}
\title{data.collapse}
\usage{
data.collapse(mrna, data)
}
\arguments{
\item{mrna}{: dataframe for mrna (follows demo)}

\item{data}{: list of data that needs to consolidated}

\item{PC_VAR_THRESH}{: The threshold for PCs to accept}

\item{pc_collapse}{: The additional consolidation function applied to selected PCs}

\item{DEBUG}{FALSE: flag to print debug statements}
}
\description{
Collapses data which has >1 column per gene using PCA
}
