% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{data.validate.patients}
\alias{data.validate.patients}
\title{data.validate.patients}
\usage{
data.validate.patients(mrna,outcome,data)

For each dataset the patient vector is assumed to be the rownames.
}
\arguments{
\item{mrna}{:mrna data (follow convention of demo_mrna)}

\item{data}{:list of other dataframes (follow demo conventions)}

\item{DEBUG}{FALSE: flag to print debug statements}

\item{outcome:}{outcome data (follow demo conventions)}
}
\value{
boolean: whether this set of iBAG data is valid for the patients only
}
\description{
Validates mrna, outcome, and upstream data to ensure patients match.
}
\details{
This function checks:
\itemize{
\item patients match across the list of datasets in data and mrna
\item patients in outcome are in mrna and data
}

A match across patient vector is defined as equivalent order and content.
}
