% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{data.validate}
\alias{data.validate}
\title{data.validate}
\usage{
data.validate(mrna,outcome,data)
}
\arguments{
\item{mrna}{mrna data (follow convention of demo_mrna)}

\item{outcome}{outcome data (follow demo conventions)}

\item{data}{list of other dataframes (follow demo conventions)}

\item{DEBUG}{(FALSE) flag to print debug statements}
}
\value{
(boolean) whether this set of iBAG data is valid
}
\description{
Validates mrna, outcome, and upstream data to ensure patients & genes match.
}
\details{
This function checks:
\itemize{
\item patients match across the list of datasets in data and mrna
\item patients in outcome are in mrna and data
\item genes match across mrna & data
}

A match across patient vector is defined as equivalent order and content.
A match across gene vector is defined as each gene in mrna appears at least
once in each dataset in data.
}
