% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{dataset.collapse.pc.singular}
\alias{dataset.collapse.pc.singular}
\title{dataset.collapse.pc.singular}
\usage{
data.collapse(mrna, data)
}
\arguments{
\item{mrna}{: dataframe for mrna (follows demo)}

\item{dataset}{: dataset that needs to consolidated (follows format demo)}

\item{PC_VAR_THRESH}{(0.09): The \% threshold for PCs to accept based on scale of eigenvalue (must be <1)}

\item{pc_collapse}{: The additional consolidation function applied to selected PCs. (default take the max)}

\item{DEBUG}{FALSE: flag to print debug statements}
}
\description{
Collapses data which has >1 column per gene using PCA
}
