% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_mrna}
\alias{demo_mrna}
\title{mRNA expression data for ???}
\format{
A data frame with ??? rows and ??? variables
}
\usage{
data(demo_mrna)
}
\description{
mRNA expression data for ???
}
\keyword{datasets}
