% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_outcome}
\alias{demo_outcome}
\title{outcome data for ???}
\format{
A data frame with ??? rows and ??? variables
}
\usage{
data(demo_outcome)
}
\description{
outcome data for ???
}
\keyword{datasets}
