% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dsurv}
\alias{dsurv}
\title{Survival data}
\format{
A data frame with 163 rows and 2 variables
}
\usage{
data(dsurv)
}
\description{
Survival data, measured in ???, of 163 individuals suffering from Glioblastoma.
Data obtained from ??? at ???.
}
\details{
\itemize{
\item X. TCGA patient ID
\item V1. Patient outcomes measured in ???
}
}
\keyword{datasets}
