% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get.data.names}
\alias{get.data.names}
\title{get.data.names}
\usage{
get.data.names(
  data.size,
  data.names,
  sep = "_",
  default.data.name = "data",
  DEBUG = FALSE
)
}
\description{
get a vector of usable names for the list of datasets.
}
