% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAG.R
\docType{package}
\name{iBAG}
\alias{iBAG}
\title{iBAG: A package for fitting the iBAG model.}
\description{
Get started with iBAG in 2 lines:
demo_data <- iBAG::demo_data()
iBAG_result <- iBAG::fit(demo_data)
summary(iBAG_result)
summary(iBAG_result$mechmodel)
}
\section{iBAG functions}{

The iBAG functions ...
}

