% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{iBAG_results}
\alias{iBAG_results}
\title{iBAG_results}
\usage{
iBAG_results$new()
}
\description{
R6 class to store and manage the iBAG results.
This does not contain a copy of the iBAG_data object.

constructor
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{iBAG_results$new()}}
\item \href{#method-get.X}{\code{iBAG_results$get.X()}}
\item \href{#method-get.Y}{\code{iBAG_results$get.Y()}}
\item \href{#method-get.SS}{\code{iBAG_results$get.SS()}}
\item \href{#method-get.beta_mean}{\code{iBAG_results$get.beta_mean()}}
\item \href{#method-get.beta_incl_prob}{\code{iBAG_results$get.beta_incl_prob()}}
\item \href{#method-get.n_patients}{\code{iBAG_results$get.n_patients()}}
\item \href{#method-get.n_genes}{\code{iBAG_results$get.n_genes()}}
\item \href{#method-get.n_data}{\code{iBAG_results$get.n_data()}}
\item \href{#method-clone}{\code{iBAG_results$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$new(
  X = matrix(0, 1, 1),
  Y = c(0),
  SS = list(c(0), c(0), c(0)),
  beta_mean = c(0),
  beta_incl_prob = c(0),
  DEBUG = FALSE,
  validate = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{(NULL): the total X matrix generated by the mech model}

\item{\code{Y}}{(NULL): the vector of outcomes that goes into the clinical model}

\item{\code{SS}}{(list(NULL,NULL)): a list of vectors representing the Sum of Squares Error from the mechanistic model.}

\item{\code{beta_mean}}{(NULL): a vector of posterior means for the coefficients from the clinical model.}

\item{\code{beta_incl_prob}}{(NULL): a vector of posterior inclusion probabilities for the beta coefficients.}

\item{\code{DEBUG}}{(FALSE): initialize object in DEBUG model.}

\item{\code{validate}}{(TRUE): validate the data coming in from the constructor (or any of the setters)}

\item{\code{...}}{: ...}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This constructs a class containing all the results from the iBAG object.
Some additional details:
\itemize{
\item SS:
\itemize{
\item The list size is always >2. The 1st element is the SST, the last element is the SSO,
the middle elements correspond to the sum of squares from the datasets.
}
\item validate:
\itemize{
\item minimum data required to validate is X,Y,SS (all produced by mechmodel)
\itemize{
\item if this data is not provided will issue warning & set validate to FALSE
}
\item n_data is derived from length(SS)-2
\item n_patients is derived as the nrow(X)
\item check that ncol(X)\%n_data == 0
\itemize{
\item if not throw error
\item if true: n_genes = ncol(X)\%/\%n_data
}
\item check that length(Y) == n_patients
\itemize{
\item if not throw error
}
\item for each item in SS
\itemize{
\item check that length(item) == n_genes
\itemize{
\item if not throw error
}
}
\item if beta_mean & beta_incl_prob are included
\itemize{
\item check that their length == ncol(X)
\itemize{
\item if not throw error
}
}
\item This does not check that the data is in a numeric form! (Check that yourself)
get.X
}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.X"></a>}}
\if{latex}{\out{\hypertarget{method-get.X}{}}}
\subsection{Method \code{get.X()}}{
returns X matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$get.X()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
matrix that should be n_patients rows and n_genes*n_data columns
get.Y
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.Y"></a>}}
\if{latex}{\out{\hypertarget{method-get.Y}{}}}
\subsection{Method \code{get.Y()}}{
returns Y vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$get.Y()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector that is n_patients long
get.SS
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.SS"></a>}}
\if{latex}{\out{\hypertarget{method-get.SS}{}}}
\subsection{Method \code{get.SS()}}{
returns the Sum of Squares mech model results, either all or some of them
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$get.SS(index = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{(NULL) can be NULL, integer, or string. If NULL returns list, otherwise returns the vector at index}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector or a list of vectors
get.beta_mean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.beta_mean"></a>}}
\if{latex}{\out{\hypertarget{method-get.beta_mean}{}}}
\subsection{Method \code{get.beta_mean()}}{
returns the vector of posterior beta means
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$get.beta_mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a vector of posterior beta means
get.beta_incl_prob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.beta_incl_prob"></a>}}
\if{latex}{\out{\hypertarget{method-get.beta_incl_prob}{}}}
\subsection{Method \code{get.beta_incl_prob()}}{
returns a vector of posterior beta inclusion probabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$get.beta_incl_prob()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a vector of posterior beta inclusion probabilities
get.n_patients
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.n_patients"></a>}}
\if{latex}{\out{\hypertarget{method-get.n_patients}{}}}
\subsection{Method \code{get.n_patients()}}{
returns the number of patients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$get.n_patients()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer of number of patients
get.genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.n_genes"></a>}}
\if{latex}{\out{\hypertarget{method-get.n_genes}{}}}
\subsection{Method \code{get.n_genes()}}{
returns the number of genes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$get.n_genes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer of number of genes
get.n_data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.n_data"></a>}}
\if{latex}{\out{\hypertarget{method-get.n_data}{}}}
\subsection{Method \code{get.n_data()}}{
returns the number of upstream datasets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$get.n_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
integer of number of upstream datasets
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_results$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
