% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechmodel.R
\name{mech.model}
\alias{mech.model}
\title{mech.model}
\usage{
mech.model(
  mrna,
  data.list,
  sep = "_",
  other.name = "other",
  default.data.name = "data",
  DEBUG = FALSE,
  ...
)
}
\arguments{
\item{mrna}{the mrna data. Look at iBAG::demo_mrna for example format}

\item{data.list}{a list of the upstream data}

\item{sep}{(_): the character value to use as a seperator}

\item{other.name}{("other"): the string to use to label residuals from mechmodel}

\item{default.data.name}{("data"): the string to use to label unnamed datasets from data.list}

\item{DEBUG}{(FALSE): debug flag}

\item{...}{extra arguments for the iBAG model}
}
\description{
This fits the mechanistic model for iBAG
}
