% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechmodel.R
\name{mech.model.single_gene_val}
\alias{mech.model.single_gene_val}
\title{mechmodel}
\usage{
mech.model.single_gene_val(mrna, data.list, DEBUG = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
mechmodel
}
