% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAGlinear.R
\name{mech_model}
\alias{mech_model}
\title{Mechanistic Model}
\usage{
mech_model(meth, mrna, cnv, verbose = F)
}
\arguments{
\item{meth}{an N (individuals) by k (genes) matrix/dataframe of methylation data (numerical values)}

\item{mrna}{an N (individuals) by k (genes) matrix/dataframe of mrna expression (numerical values)}

\item{cnv}{an N (individuals) by k (genes) matrix/dataframe of cnv data}

\item{verbose}{boolean, if True print and return debugging information}
}
\value{
obj An object containing the results.
obj$:
obj$:
}
\description{
Fit the iBAG mechanistic model on the data
}
\examples{
sim()


}
