% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{meth}
\alias{meth}
\title{Methylation data for ???}
\format{
A data frame with ??? rows and ??? variables
}
\usage{
data(meth)
}
\description{
A dataset containing the methylation data for 163 individuals for ??? genes across ??? sites.
Data obtained from ??? at ???
}
\details{
\itemize{
\item ???
}
}
\keyword{datasets}
