% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mrna}
\alias{mrna}
\title{Gene expression data}
\format{
A data frame with 163 rows and ??? variables
}
\usage{
data(mrna)
}
\description{
A dataset containing the mrna expression data for 163 individuals for ??? genes across ??? sites.
Data obtained from ??? at ???
}
\details{
\itemize{
\item ???
}
}
\keyword{datasets}
