% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAGlinear.R
\name{proc_user_data}
\alias{proc_user_data}
\title{Process user data}
\usage{
proc_user_data(meth, cnv, expr, out, verbose = F)
}
\arguments{
\item{meth}{an N (individuals) by k (genes) matrix/dataframe of methylation data (numerical values)}

\item{cnv}{an N (individuals) by k (genes) matrix/dataframe of cnv data}

\item{expr}{an N (individuals) by k (genes) matrix/dataframe of mrna expression (numerical values)}

\item{out}{a N length vector of outcomes or Nx2 matrix/dataframe of patient id (string) & outcome (numeric)}

\item{verbose}{boolean, if True print and return debugging information}
}
\value{
obj An object containing the results.
obj$meth:
obj$cnv:
obj$expr:
obj$outcome:
obj$gene_names:
obj$patient_names:
obj$N: The number of individuals
obj$K: The number of genes
obj$
}
\description{
Check data constraints (Throw Warnings/Stop if not met):
\enumerate{
\item Null checks
\item N matches for all data submitted by user (meth, cnv, expr, outcome)
\item Genes match across all platforms (meth, cnv, expr)
\item Genes/Individual ratio <= 5. User cannot supply too many genes.
return a standardized set of variables to apply the iBAG procedure on
}
}
\examples{
()

}
