% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{data.validate}
\alias{data.validate}
\title{data.validate}
\usage{
data.validate(mrna,outcome,data)
}
\arguments{
\item{mrna}{mrna data (follow convention of demo_mrna)}

\item{outcome}{outcome data (follow demo conventions)}

\item{data}{list of other dataframes (follow demo conventions)}

\item{sep}{("_"): the character seperator to use to differentiate mutliple columns of data per gene}

\item{DEBUG}{(FALSE) flag to print debug statements}
}
\value{
(boolean) whether this set of iBAG data is valid
}
\description{
Validates mrna, outcome, and upstream data to ensure patients & genes match.
}
\details{
Runs the data.validate.patients & data.validate.genes subroutines. Consult them directly for more information
}
