% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_cnv}
\alias{demo_cnv}
\title{demo_cnv}
\format{
A data frame with 770 rows and 1219 variables
}
\usage{
data(demo_cnv)
}
\description{
demo_cnv
}
\details{
cnv data for TCGA-BRCA, note we filter out to only 1218 genes to keep the dataset "small"
We already condense the cnv to single number summaries per gene
The 1st column refers to the patient ids
}
\keyword{datasets}
