% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_meth}
\alias{demo_meth}
\title{Methylation data for ???}
\format{
A data frame with ??? rows and ??? variables
}
\usage{
data(demo_meth)
}
\description{
Methylation data for ???
}
\keyword{datasets}
