% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_meth}
\alias{demo_meth}
\title{demo_meth}
\format{
A data frame with 770 rows and 1218 variables
}
\source{
https://portal.gdc.cancer.gov/
}
\usage{
data(demo_meth)
}
\description{
demo_meth
}
\details{
Methylation data for TCGA-BRCA, note we filter out to only 1218 genes to keep the dataset "small".
rownames are the patient ids.
colnames are the gene names.
We also collapse the multiple probe information from TCGA using dataset.collapse.pc.singular so we only get 1 column per gene.
}
\keyword{datasets}
