% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_mrna}
\alias{demo_mrna}
\title{demo_mrna}
\format{
A data frame with 770 rows and 1218 variables
}
\usage{
data(demo_mrna)
}
\description{
demo_mrna
}
\details{
mRNA expression data for TCGA-BRCA, note we filter out to only 1218 genes to keep the dataset "small".
rownames are the patient ids.
colnames are the gene names.
}
\keyword{datasets}
