% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_outcome}
\alias{demo_outcome}
\title{demo_outcome}
\format{
A data frame with 770 rows and 1 variables
}
\source{
https://portal.gdc.cancer.gov/
}
\usage{
data(demo_outcome)
}
\description{
demo_outcome
}
\details{
EREG-mRNA stemness index outcome data for TCGA-BRCA. Consult https://doi.org/10.1016/j.cell.2018.03.034 for more details.
}
\keyword{datasets}
