% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dsurv}
\alias{dsurv}
\title{Survival data of 163 individuals.
Data obtained from ??? at ???}
\format{
A data frame with 163 rows and 2 variables
}
\usage{
data(dsurv)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds. The variables are as follows:
}
\details{
\itemize{
\item X. TCGA patient ID
\item V1. Patient outcomes measured in ???
}
}
\keyword{datasets}
