% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAGlinear.R
\name{iBAG}
\alias{iBAG}
\title{Fit the iBAG framework on the data}
\usage{
iBAG(gene_expr, cnv, meth, out, iBAGtype = "linear", fast = T, verbose = F)
}
\arguments{
\item{gene_expr}{an N (individuals) by k (genes) matrix/dataframe of mrna expression (numeric)}

\item{cnv}{an N (individuals) by k (genes) matrix/dataframe of cnv data (numeric)}

\item{meth}{an N (individuals) by k (genes) matrix/dataframe of methylation data (numeric)}

\item{out}{a N length vector of outcomes or Nx2 matrix/dataframe of patient id (string) & outcome (numeric)}

\item{iBAGtype}{c("linear","discrete","survival")}

\item{fast}{boolean, if True use the EMVS approach, otherwise run full MCMC}

\item{verbose}{boolean, if True print and return debugging information}
}
\value{
obj An object containing the results.
obj$mech:
obj$clinical:
}
\description{
Fit the iBAG framework on the data
}
\examples{
sim()

}
