% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{iBAG_data}
\alias{iBAG_data}
\title{iBAG_data}
\description{
iBAG_data

iBAG_data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{iBAG_data$new()}}
\item \href{#method-get.meth}{\code{iBAG_data$get.meth()}}
\item \href{#method-get.cnv}{\code{iBAG_data$get.cnv()}}
\item \href{#method-get.mrna}{\code{iBAG_data$get.mrna()}}
\item \href{#method-get.outcome}{\code{iBAG_data$get.outcome()}}
\item \href{#method-clone}{\code{iBAG_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_data$new(DEBUG = FALSE, print_status = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DEBUG}}{boolean: initialize object in DEBUG mode}

\item{\code{print_status}}{boolean: initialize object to print (to stdout) everything that it does}

\item{\code{...}}{: ...}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.meth"></a>}}
\if{latex}{\out{\hypertarget{method-get.meth}{}}}
\subsection{Method \code{get.meth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_data$get.meth()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.cnv"></a>}}
\if{latex}{\out{\hypertarget{method-get.cnv}{}}}
\subsection{Method \code{get.cnv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_data$get.cnv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.mrna"></a>}}
\if{latex}{\out{\hypertarget{method-get.mrna}{}}}
\subsection{Method \code{get.mrna()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_data$get.mrna()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.outcome"></a>}}
\if{latex}{\out{\hypertarget{method-get.outcome}{}}}
\subsection{Method \code{get.outcome()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_data$get.outcome()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iBAG_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
